+++ lib/pay/braintree/webhooks/subscription_canceled.rb
@@ -8,15 +8,7 @@ module Pay
           subscription = event.subscription
           return if subscription.nil?
 
-          pay_subscription = Pay::Subscription.find_by_processor_and_id(:braintree, subscription.id)
-          return unless pay_subscription.present?
-
-          ends_at = Time.current
-          pay_subscription.update!(
-            status: :canceled,
-            trial_ends_at: (ends_at if pay_subscription.trial_ends_at?),
-            ends_at: ends_at
-          )
+          Pay::Braintree::Subscription.sync(subscription.id)
         end
       end
     end