+++ lib/pay/braintree/webhooks/subscription_expired.rb
@@ -8,10 +8,7 @@ module Pay
           subscription = event.subscription
           return if subscription.nil?
 
-          pay_subscription = Pay::Subscription.find_by_processor_and_id(:braintree, subscription.id)
-          return unless pay_subscription.present?
-
-          pay_subscription.update!(ends_at: Time.current, status: :canceled)
+          Pay::Braintree::Subscription.sync(subscription.id)
         end
       end
     end